Fix armor not taking damage or being enchantable

+++ common/src/main/java/com/grim3212/assorted/tools/common/item/configurable/ConfigurableShovelItem.java
@@ -35,8 +35,8 @@ public class ConfigurableShovelItem extends ShovelItem implements ITiered, IItem
     public Multimap<Attribute, AttributeModifier> getDefaultAttributeModifiers(EquipmentSlot slot) {
         if (attribs == null) {
             Builder<Attribute, AttributeModifier> builder = ImmutableMultimap.builder();
-            builder.put(Attributes.ATTACK_DAMAGE, new AttributeModifier(BASE_ATTACK_DAMAGE_UUID, "Tool modifier", (double) this.getAttackDamage(), AttributeModifier.Operation.ADDITION));
-            builder.put(Attributes.ATTACK_SPEED, new AttributeModifier(BASE_ATTACK_SPEED_UUID, "Tool modifier", (double) SHOVEL_SPEED, AttributeModifier.Operation.ADDITION));
+            builder.put(Attributes.ATTACK_DAMAGE, new AttributeModifier(BASE_ATTACK_DAMAGE_UUID, "Tool modifier", this.getAttackDamage(), AttributeModifier.Operation.ADDITION));
+            builder.put(Attributes.ATTACK_SPEED, new AttributeModifier(BASE_ATTACK_SPEED_UUID, "Tool modifier", SHOVEL_SPEED, AttributeModifier.Operation.ADDITION));
             this.attribs = builder.build();
         }
